support for revved files with suffix naming instead of prefix naming (foo.7483.js instead of 7483.foo.js)
switched from mocha to grunt-mocha-cli to enable testing on Windows. also results in less config and prettier colors.  normalized file paths in test code to make tests pass on Windows.

fixed a globbing bug that only occurs on Mac OS when the pattern starts with #.  hooray for the existing test which caught this!